requirements.txt
Flask==3.0.3
tensorflow==2.14.0
keras==2.14.0
numpy==1.26.4
pillow==10.2.0
opencv-python==4.9.0.80
